% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_from_yahoo}
\alias{get_from_yahoo}
\title{Get the data from Yahoo Japan}
\usage{
get_from_yahoo(code, start_date, end_date, frequency)
}
\arguments{
\item{code}{Dataset code on Quandl specified as a string or an array of strings.}

\item{start_date}{Use to truncate data by start date in 'yyyy-mm-dd' format.}

\item{end_date}{Use to truncate data by end date in 'yyyy-mm-dd' format.}

\item{collapse}{Collapse frequency of Data.}
}
\value{
data
}
\description{
Get the data from Yahoo Japan
}

