% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{rfinancej}
\alias{rfinancej}
\title{Get financial data in Japan}
\usage{
rfinancej(code, type, start_date = Sys.Date() - 365, end_date = Sys.Date(),
  frequency = "daily", src = "yahoo")
}
\arguments{
\item{code}{Dataset code on different type of data source specified as a string or an array of strings.}

\item{type}{Type of data returned specified as string. Can be 'data.frame', 'xts'.}

\item{start_date}{Use to truncate data by start date in 'yyyy-mm-dd' format or Date object.}

\item{end_date}{Use to truncate data by end date in 'yyyy-mm-dd' format  or Date object.}

\item{frequency}{Frequency of Data(daily, weekly, mothly, annual)}

\item{src}{data source Yahoo Japan! Finance,}
}
\value{
data
}
\description{
Get financial data in Japan
}

